Remove some dependencies on k0s in inttests
This commit is the first of many to remove the dependency of
github.com/k0sproject/k0s.

So far, we're removing github.com/k0sproject/k0s/inttest/common in every
test that doesn't depend on FootlooseSuite.

There have been a couple modifications due to deprecation and due to
incompatibilities with macOS.

Signed-off-by: Juan-Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>