Fix PVC bindind when Cluster PVC name is given
When the cluster persistence was of type PVC and the cluster name was given by the user, the statefulset related to the k0smotron Cluster could not create the pods because the init container was using the wrong mount name. It was using the name kmc-<cluster.name> instead of the one configured by the user.

Signed-off-by: Adrian Pedriza <adripedriza@gmail.com>